Workaround: WatchOS ARM stack wrong after exception
Force watchOS to adjust the stack in function epilog using the frame
pointer.

This is a workaround for a bug elsewhere that creates an incorrect stack
after an exception is caught.  Not sure exactly where the bug is.  It
happens when stack adjustment is needed around a function that throws
and stack is not fixed in landing pad.  Stack is never corrected, leaks,
and registers are incorrectly restored in epilog if stack not restored
from frame pointer.  Can partially create problem with clang++ by using
VLA and variadic call.  It creates the same stack leak after caught
exception but it then falls back to using frame pointer in epilog so no
crash (yet).